some more u32

+++ expander_compiler/tests/zkcuda/zkcuda_examples.rs
@@ -48,7 +48,7 @@ fn zkcuda_test<C: Config, P: ProvingSystem<C>>() {
     call_kernel!(ctx, kernel_add_16, 1, b, mut c).unwrap();
     let c = c.reshape(&[]);
     let result: CircuitField<C> = ctx.copy_to_host(c);
-    assert_eq!(result, CircuitField::<C>::from(32 * 33 / 2));
+    assert_eq!(result, CircuitField::<C>::from(32 * 33 / 2 as u32));
 
     let computation_graph = ctx.compile_computation_graph().unwrap();
     ctx.solve_witness().unwrap();
@@ -256,7 +256,7 @@ fn zkcuda_to_binary() {
     let kernel: KernelPrimitive<M31Config> = compile_convert_to_binary().unwrap();
     let mut ctx: Context<M31Config, _> = Context::new(hint_registry);
 
-    let a = M31::from(0x55);
+    let a = M31::from(0x55 as u32);
     let a = ctx.copy_to_device(&a);
     let a = a.reshape(&[1]);
     let mut b: DeviceMemoryHandle = None;
@@ -266,14 +266,14 @@ fn zkcuda_to_binary() {
     assert_eq!(
         result,
         vec![
-            M31::from(1),
-            M31::from(0),
-            M31::from(1),
-            M31::from(0),
-            M31::from(1),
-            M31::from(0),
-            M31::from(1),
-            M31::from(0)
+            M31::from(1u32),
+            M31::from(0u32),
+            M31::from(1u32),
+            M31::from(0u32),
+            M31::from(1u32),
+            M31::from(0u32),
+            M31::from(1u32),
+            M31::from(0u32)
         ]
     );
